split convert testing files
